+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Improves Error Deserialization to dynamically handle edge cases that have a bad format
+
 ## v4.13.0 (2023-04-04)
 
 - Adds `get_next_page` function to each object which retrieves the next page of a collection when the `has_more` key is present in the response